requirements.txt
streamlit
openai
pydantic
python-dotenv
sentence-transformers
numpy
pandas
faiss-cpu
sqlalchemy
email-validator
python-dateutil
PyPDF2
